Create separate file for tau related JS functionality
