+++ caseworker/tau/templates/tau/home.html
@@ -120,3 +120,8 @@
     </main>
 </div>
 {% endblock %}
+
+{% block javascript %}
+    {{ block.super }}
+    <script src="{% static 'tau.js' %}" type="text/javascript"></script>
+{% endblock %}
