+++ package.json
@@ -5,9 +5,9 @@
   "private": true,
   "scripts": {
     "build_exporter": "parcel build ./exporter/assets/javascripts/main.js ./exporter/assets/javascripts/cookie-policy-form.js --dist-dir exporter/assets/built --public-url /assets/",
-    "build_caseworker": "parcel build ./caseworker/assets/javascripts/{main,head,cookie-policy-form,search-cases,review-good,search-products}.js --dist-dir caseworker/assets/built --public-url /assets/",
+    "build_caseworker": "parcel build ./caseworker/assets/javascripts/{main,head,cookie-policy-form,search-cases,review-good,search-products,tau}.js --dist-dir caseworker/assets/built --public-url /assets/",
     "watch_exporter": "parcel watch ./exporter/assets/javascripts/{main,head}.js ./exporter/assets/javascripts/cookie-policy-form.js --dist-dir exporter/assets/built --public-url /assets/",
-    "watch_caseworker": "parcel watch ./caseworker/assets/javascripts/{main,head,cookie-policy-form,search-cases,review-good,search-products}.js --dist-dir caseworker/assets/built --public-url /assets/",
+    "watch_caseworker": "parcel watch ./caseworker/assets/javascripts/{main,head,cookie-policy-form,search-cases,review-good,search-products,tau}.js --dist-dir caseworker/assets/built --public-url /assets/",
     "watch": "npm-run-all --parallel watch_exporter watch_caseworker",
     "build": "export NODE_ENV=production && npm-run-all --serial build_exporter build_caseworker",
     "test:e2e": "cypress run --browser chrome",
